Bump spades back to 3.13.0 per Unicycler release notes:
Nothing major here - just making a release to include a handful of fixes that were on the main branch but not yet in a release.

The most significant one is that Unicycler now enforces that SPAdes cannot be a version later than v3.13.0. This is because after that version, SPAdes stopped saving intermediate files that Unicycler needs. Unicycler users that had later versions of SPAdes were getting suboptimal assemblies.