Fixes for large numbers of queued jobs.
Where we choose an output name that is the same as a table column, but  compute
a value the index on that field cannot be used unless the sort clause
explicitly names the table and field.

In our main scheduler loop query, Order by ID instead of submission
time. With a large number of queued jobs and no index on both
state_code and submit time, mysql chooses the index it has
(state_code) and sorts by time.  ID is a good proxy for submit time.